chore: move rejected plugin to separate page
