+++ projects/demo/src/pages/kit/plugins/examples/1-reject/component.ts
