+++ projects/demo/src/pages/kit/plugins/examples/1-reject/index.ts
