+++ projects/demo/src/pages/kit/plugins/examples/1-reject/mask.ts
@@ -1,0 +1,13 @@
+import type {MaskitoOptions} from '@maskito/core';
+import {maskitoRejectEvent} from '@maskito/kit';
+
+const maskitoOptions: MaskitoOptions = {
+    mask: /^\d{0,3}$/,
+    plugins: [
+        // This plugin dispatches custom event `maskitoReject` when a character that the
+        // user has entered is rejected by the mask. You can use it to visualize rejection.
+        maskitoRejectEvent,
+    ],
+};
+
+export default maskitoOptions;
