+++ projects/demo/src/pages/kit/plugins/kit-plugins-doc.template.html
@@ -1,0 +1,14 @@
+<tui-doc-page
+    header="Plugins"
+    package="KIT"
+>
+    <tui-doc-example
+        id="reject"
+        heading="Visualize rejected characters"
+        [content]="rejectExample"
+    >
+        <plugins-reject-doc-example-1 />
+    </tui-doc-example>
+
+    <!-- TODO-->
+</tui-doc-page>
